refactor: resolve PR feedback and architectural blockers
- Remove production debugPrint statements from GlassDragBuilder.
- Fix circular dependency by moving InteractionNotification to a shared internal renderer file.
- Centralize default sheet settings into a shared constants file for GlassSheet and GlassModalSheet.
- Restore technical documentation and coordinate conversion comments in glass_glow.dart.
