+++ lib/widgets/overlays/shared/glass_modal_sheet_state.dart
@@ -651,7 +651,7 @@ class _GlassModalSheetState extends State<GlassModalSheet>
     final baseSettings = GlassThemeHelpers.resolveSettings(
       context,
       explicit: widget.settings,
-      fallback: _kDefaultSheetSettings,
+      fallback: kDefaultSheetSettings,
     );
 
     final focusBridge = Focus(
